test(server): group workflow action suites by action and driver
